Björn's patch
The new don't-stop-the-world text patching on change only one RISC-V
instruction:

  |  -8: &ftrace_ops of the associated tracer function.
  | <ftrace enable>:
  |   0: auipc  t0, hi(ftrace_caller)
  |   4: jalr   t0, lo(ftrace_caller)
  |
  |  -8: &ftrace_nop_ops
  | <ftrace disable>:
  |   0: auipc  t0, hi(ftrace_caller)
  |   4: nop

This means that f+0x0 is fixed, and should not be claimed by ftrace,
e.g. kprobe should be able to put a probe in f+0x0.

In fact, the kprobe scheme is used by many tests in the BPF selftest
suite.

TODO: Is the FENTRY_RA_OFFSET usage correct for ftrace graph

Andy, feel free to fold this into the series!

Signed-off-by: Björn Töpel <bjorn@rivosinc.com>